Refactor things into a library model to allow for testing
To properly run tests on the project, this commit refactors code into a library model. This involves main.rs importing the library instead of declaring the modules itself. This is required for cargo to run docs tests. Additionally, a new library target gets created for cargo to run tests on. This is not meant to be actually imported separately and is only used in this project.